Refactor to clean architecture; add services, DTOs, repos

+++ GardenShop.Api.Application/Abstractions/IRepositories/ICategoryRepository.cs
@@ -5,9 +5,9 @@ namespace GardenShop.Application.Abstractions.Persistence
 	public interface ICategoryRepository : IRepository<Category>
 	{
 		Task<List<Category>> GetAllAsync(CancellationToken ct);
-		Task<Category?> GetByIdWithProductsAsync(int id, CancellationToken ct);
+		Task<Category?> GetByIdWithProductsAsync(long id, CancellationToken ct);
 
-		Task<bool> NameExistsAsync(string name, int? exceptId, CancellationToken ct);
-		Task<bool> IsUsedByProductsAsync(int categoryId, CancellationToken ct);
+		Task<bool> NameExistsAsync(string name, long? exceptId, CancellationToken ct);
+		Task<bool> IsUsedByProductsAsync(long categoryId, CancellationToken ct);
 	}
 }
