+++ GardenShop.Api.Domain/Entities/OrderItem.cs
@@ -2,10 +2,10 @@
 {
 	public class OrderItem
 	{
-		public int OrderId { get; set; }
+		public long OrderId { get; set; }
 		public Order Order { get; set; } = null!;
 
-		public int ProductId { get; set; }
+		public long ProductId { get; set; }
 		public Product Product { get; set; } = null!;
 
 		public required string ProductNameSnapshot { get; set; }
